app/requirements.txt
flask==2.3.3
werkzeug==2.3.7
cryptography==41.0.0
psycopg2-binary==2.9.9